ensure permissions are set up correctly
